Add Gcloud::Upload
Move resumable_threshold out of Storage into its own namespace.
This allows both Storage and BigQuery to use the setting.
