+++ lib/gcloud/storage.rb
@@ -338,26 +338,5 @@ module Gcloud
   #   file.acl.public!
   #
   module Storage
-    ##
-    # Retrieve resumable threshold.
-    # If uploads are larger in size than this value then
-    # resumable uploads are used.
-    #
-    # The default value is 5 MiB (5,000,000 bytes).
-    def self.resumable_threshold
-      @@resumable_threshold
-    end
-
-    ##
-    # Sets a new resumable threshold value.
-    def self.resumable_threshold= new_resumable_threshold
-      # rubocop:disable Style/ClassVars
-      # Disabled rubocop because this is the best option.
-      @@resumable_threshold = new_resumable_threshold.to_i
-      # rubocop:enable Style/ClassVars
-    end
-
-    # Set the default threshold to 5 MiB.
-    self.resumable_threshold = 5_000_000
   end
 end
